lms-backend/requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
Automat==24.8.1
av==14.4.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.8.2
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
ConfigArgParse==1.7.1
constantly==23.10.4
contourpy==1.3.2
cors==1.0.1
ctranslate2==4.6.0
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
faster-whisper==1.1.1
filelock==3.18.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
future==1.0.0
gaze-tracking==0.0.1
gevent==24.11.1
geventhttpclient==2.3.4
googletrans
greenlet==3.1.1
gTTS==2.5.4
gunicorn==21.2.0
h11==0.16.0
h2==3.2.0
hf-xet==1.1.3
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx
huggingface-hub==0.32.4
humanfriendly==10.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
iniconfig==2.0.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
locust==2.37.10
locust-cloud==1.23.2
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
nltk==3.8.1
numpy==1.26.4
onnxruntime==1.22.0
openai==1.84.0
opencv-contrib-python==4.11.0.86
opencv-python==4.8.0.76
opt_einsum==3.4.0
packaging==24.2
pandas==2.1.4
Pillow==10.0.0
platformdirs==4.3.8
pluggy==1.5.0
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyjson==1.4.1
pymongo==4.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pypickle==1.1.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.2
python-jose==3.3.0
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rfc3986==1.5.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.0
scipy==1.15.3
sentencepiece==0.2.0
setuptools==75.8.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sympy==1.14.0
threadpoolctl==3.6.0
tldextract==5.1.3
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
Twisted==24.11.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
zope.event==5.0
zope.interface==7.2
